requirements.txt
aiogram==3.0.0
sulguk==0.5.0
pydantic-settings==2.0.3
pyaml==23.9.6
structlog~=23.1.0
# asyncio speedup
uvloop==0.17.0; sys_platform == 'linux'
